secretreader: Refactor secretreader into single file
